Skip the install tests in travis.

+++ examples_tests/tests.py
@@ -26,6 +26,7 @@ import testscenarios
 
 
 logger = logging.getLogger(__name__)
+config = {}
 
 
 def _get_ubuntu_version():
@@ -33,10 +34,51 @@ def _get_ubuntu_version():
         ['lsb_release', '-cs']).decode('utf8').strip()
 
 
+def _wait_for_ssh(ip, port):
+    logger.info('Waiting for ssh to be enable in the testbed...')
+    timeout = 300
+    sleep = 10
+    while (timeout > 0):
+        try:
+            _run_command_through_ssh(['echo', 'testing ssh'])
+            break
+        except subprocess.CalledProcessError:
+            if timeout <= 0:
+                logger.error('Timed out waiting for ssh in the testbed.')
+                raise
+            else:
+                time.sleep(sleep)
+                timeout -= sleep
+
+
+def _run_command_through_ssh(ip, port, command):
+    ssh_command = ['ssh', 'ubuntu@' + ip, '-p', port]
+    ssh_command.extend(_get_ssh_options())
+    ssh_command.extend(command)
+    return subprocess.check_output(ssh_command).decode("utf-8")
+
+
+def _get_ssh_options():
+    return [
+        '-o', 'UserKnownHostsFile=/dev/null',
+        '-o', 'StrictHostKeyChecking=no',
+        '-i', os.path.join(os.getenv('HOME'), '.ssh', 'id_rsa')
+    ]
+
+
+def _copy_file_through_ssh(ip, port, user, local_file_path, remote_file_path):
+    scp_command = ['scp', '-P', self.testbed_port]
+    scp_command.extend(_get_ssh_options())
+    scp_command.extend([
+        local_file_path, '{}@{}:{}'.format(user, ip, remote_file_path)])
+    subprocess.check_call(scp_command)
+
+
 class TestSnapcraftExamples(testscenarios.TestWithScenarios):
 
     testbed_ip = 'localhost'
     testbed_port = '8022'
+    vm_process = None
 
     scenarios = [
         ('downloader-with-wiki-parts', {
@@ -114,58 +156,35 @@ class TestSnapcraftExamples(testscenarios.TestWithScenarios):
     @classmethod
     def setUpClass(cls):
         cls.temp_dir = tempfile.mkdtemp()
-        cls.image_path = os.path.join(cls.temp_dir, 'snappy.img')
-        logger.info('Creating a snappy image to run the tests.')
-        subprocess.check_call(
-            ['sudo', 'ubuntu-device-flash', '--verbose',
-             'core', '15.04', '--channel', 'stable',
-             '--output', cls.image_path, '--developer-mode'])
-        logger.info('Running the snappy image in a virtual machine.')
-        system = subprocess.check_output(['uname', '-m']).strip().decode('utf8')
-        qemu_command = ('qemu-system-' + system +
-            ' -m 512 -nographic -net user -net nic,model=virtio' +
-            ' -drive file=' + cls.image_path +
-            ',if=virtio -redir tcp:{}::22'.format(cls.testbed_port) +
-            ' -monitor none -serial none')
-        cls.vm_process = subprocess.Popen(qemu_command, shell=True)
+        global config
+        if not config.get('skip-install', False):
+            logger.info('Creating a snappy image to run the tests.')
+            cls.image_path = os.path.join(cls.temp_dir, 'snappy.img')
+            subprocess.check_call(
+                ['sudo', 'ubuntu-device-flash', '--verbose',
+                'core', '15.04', '--channel', 'stable',
+                '--output', cls.image_path, '--developer-mode'])
+            logger.info('Running the snappy image in a virtual machine.')
+            system = subprocess.check_output(
+                ['uname', '-m']).strip().decode('utf8')
+            qemu_command = ('qemu-system-' + system +
+                ' -enable-kvm' +
+                ' -m 512 -nographic -net user -net nic,model=virtio' +
+                ' -drive file=' + cls.image_path +
+                ',if=virtio -redir tcp:{}::22'.format(cls.testbed_port) +
+                ' -monitor none -serial none')
+            cls.vm_process = subprocess.Popen(qemu_command, shell=True)
+            _wait_for_ssh(cls.testebed_ip, cls.testbed_port)
 
     @classmethod
     def tearDownClass(cls):
-        cls.vm_process.kill()
+        if cls.vm_process:
+            cls.vm_process.kill()
         shutil.rmtree(cls.temp_dir)
 
-    def setUp(self):
-        super().setUp()
-        self._wait_for_ssh()
-
-    def _wait_for_ssh(self):
-        logger.info('Waiting for ssh to be enable in the testbed...')
-        timeout = 300
-        sleep = 10
-        while (timeout > 0):
-            try:
-                self.run_command_through_ssh(['echo', 'testing ssh'])
-                break
-            except subprocess.CalledProcessError:
-                if timeout <= 0:
-                    logger.error('Timed out waiting for ssh in the testbed.')
-                    raise
-                else:
-                    time.sleep(sleep)
-                timeout -= sleep
-
     def run_command_through_ssh(self, command):
-        ssh_command = ['ssh', 'ubuntu@' + self.testbed_ip, '-p', self.testbed_port]
-        ssh_command.extend(self._get_ssh_options())
-        ssh_command.extend(command)
-        return subprocess.check_output(ssh_command).decode("utf-8")
-
-    def _get_ssh_options(self):
-        return [
-            '-o', 'UserKnownHostsFile=/dev/null',
-            '-o', 'StrictHostKeyChecking=no',
-            '-i', os.path.join(os.getenv('HOME'), '.ssh', 'id_rsa')
-        ]
+        return _run_command_through_ssh(
+            command, self.testebed_ip, self.testbed_port)
 
     def build_snap(self, project_dir):
         snapcraft = os.path.join(os.getcwd(), 'bin/snapcraft')
@@ -173,10 +192,9 @@ class TestSnapcraftExamples(testscenarios.TestWithScenarios):
         subprocess.check_call(snapcraft, cwd=project_dir)
 
     def copy_snap_to_testbed(self, snap_local_path):
-        scp_command = ['scp', '-P', self.testbed_port]
-        scp_command.extend(self._get_ssh_options())
-        scp_command.extend([snap_local_path, 'ubuntu@localhost:/home/ubuntu'])
-        subprocess.check_call(scp_command)
+        _copy_file_through_ssh(
+            self.testbed_ip, self.testbed_port, 'ubuntu',
+            snap_local_path, '/home/ubuntu/')
 
     def delete_snap_from_testbed(self, snap_file_name):
         self.run_command_through_ssh(
@@ -194,19 +212,24 @@ class TestSnapcraftExamples(testscenarios.TestWithScenarios):
             self.skipTest('qmldemo is not supported on trusty.')
 
         example_dir = os.path.join('examples', self.dir)
+        # Build snap will raise an exception in case of error.
         self.build_snap(example_dir)
-        snap_file_name = '{}_{}_amd64.snap'.format(self.name, self.version)
-        self.copy_snap_to_testbed(os.path.join(example_dir, snap_file_name))
-        self.addCleanup(self.delete_snap_from_testbed, snap_file_name)
-        self.install_snap(snap_file_name)
-        self.addCleanup(self.remove_snap, self.name)
-        if getattr(self, 'internal_tests_commads', None):
-            for command, expected_result in self.internal_tests_commands:
-                with self.subTest(command):
-                    output = self.run_command_through_ssh(command.split(' '))
-                    self.assertEqual(output, expected_result)
-        if getattr(self, 'external_tests_commands', None):
-            for command, expected_result in self.external_tests_commands:
-                with self.subTest(command):
-                    output = subprocess.check_output(command.split(' '))
-                    self.assertEqual(output, expected_result)
+
+        if not config.get('skip-install', False):
+            snap_file_name = '{}_{}_amd64.snap'.format(self.name, self.version)
+            self.copy_snap_to_testbed(os.path.join(example_dir, snap_file_name))
+            self.addCleanup(self.delete_snap_from_testbed, snap_file_name)
+            self.install_snap(snap_file_name)
+            self.addCleanup(self.remove_snap, self.name)
+
+            if getattr(self, 'internal_tests_commads', None):
+                for command, expected_result in self.internal_tests_commands:
+                    with self.subTest(command):
+                        output = self.run_command_through_ssh(
+                            command.split(' '))
+                        self.assertEqual(output, expected_result)
+            if getattr(self, 'external_tests_commands', None):
+                for command, expected_result in self.external_tests_commands:
+                    with self.subTest(command):
+                        output = subprocess.check_output(command.split(' '))
+                        self.assertEqual(output, expected_result)
